1.0.0-alpha 功能合入 (#14)
yapi导入支持required以及覆盖
用例编辑功能优化
添加e2e测试集
修改gitlab未遍历所有项目导致查询不到分支
数据库分布式锁优化
新增模版插件
修改复制用例问题
支持api生成模版，以及模版覆盖以及批量操作
用例compare支持参数
生成二方包优化

+++ windy-feature/src/main/java/com/zj/feature/service/TestCaseService.java
@@ -1,13 +1,20 @@
 package com.zj.feature.service;
 
 import com.baomidou.mybatisplus.core.metadata.IPage;
-import com.zj.common.generate.UniqueIdService;
+import com.zj.common.uuid.UniqueIdService;
 import com.zj.common.model.PageSize;
+import com.zj.domain.entity.dto.feature.ExecutePointDto;
+import com.zj.domain.entity.dto.feature.FeatureInfoDto;
 import com.zj.domain.entity.dto.feature.TestCaseDto;
+import com.zj.domain.repository.feature.IExecutePointRepository;
+import com.zj.domain.repository.feature.IFeatureRepository;
 import com.zj.domain.repository.feature.ITestCaseRepository;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.stereotype.Service;
-import org.springframework.util.CollectionUtils;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * @author guyuelan
@@ -17,25 +24,39 @@ import org.springframework.util.CollectionUtils;
 public class TestCaseService {
   private final UniqueIdService uniqueIdService;
   private final ITestCaseRepository testCaseRepository;
+  private final IFeatureRepository featureRepository;
+  private final IExecutePointRepository executePointRepository;
 
-  public TestCaseService(UniqueIdService uniqueIdService, ITestCaseRepository testCaseRepository) {
+  public TestCaseService(UniqueIdService uniqueIdService, ITestCaseRepository testCaseRepository, IFeatureRepository featureRepository, IExecutePointRepository executePointRepository) {
     this.uniqueIdService = uniqueIdService;
     this.testCaseRepository = testCaseRepository;
+    this.featureRepository = featureRepository;
+    this.executePointRepository = executePointRepository;
   }
 
-  public PageSize<TestCaseDto> getTestCaseList(String serviceId, Integer page, Integer pageSize) {
-    IPage<TestCaseDto> pageObj = testCaseRepository.getCasePage(serviceId, page, pageSize);
-    List<TestCaseDto> records = pageObj.getRecords();
+  public PageSize<TestCaseDto> getE2ECases(Integer page, Integer pageSize) {
+    IPage<TestCaseDto> pageObj = testCaseRepository.getE2ECasesPage(page, pageSize);
+    return convertPageSize(pageObj);
+  }
+
+  public PageSize<TestCaseDto> convertPageSize(IPage<TestCaseDto> page){
+    List<TestCaseDto> records = page.getRecords();
     PageSize<TestCaseDto> dtoPageSize = new PageSize<>();
     if (CollectionUtils.isEmpty(records)) {
       dtoPageSize.setTotal(0);
       return dtoPageSize;
     }
 
-    long total = pageObj.getTotal();
+
+    long total = page.getTotal();
     dtoPageSize.setTotal(total);
     dtoPageSize.setData(records);
     return dtoPageSize;
+  }
+
+  public PageSize<TestCaseDto> getTestCaseList(String serviceId, Integer page, Integer pageSize) {
+    IPage<TestCaseDto> pageObj = testCaseRepository.getCasePage(serviceId, page, pageSize);
+    return convertPageSize(pageObj);
   }
 
   public String createTestCase(TestCaseDto testCaseDto) {
@@ -55,7 +76,26 @@ public class TestCaseService {
     return testCaseRepository.getTestCaseById(caseId);
   }
 
+  @Transactional
   public Boolean deleteTestCase(String caseId) {
-    return testCaseRepository.deleteTestCase(caseId);
+    List<FeatureInfoDto> featureList = featureRepository.queryFeatureList(caseId);
+    if (CollectionUtils.isEmpty(featureList)) {
+      return testCaseRepository.deleteTestCase(caseId);
+    }
+
+    List<String> featureIds = featureList.stream().map(FeatureInfoDto::getFeatureId).collect(Collectors.toList());
+    List<ExecutePointDto> executePoints = executePointRepository.getPointsByFeatureIds(featureIds);
+
+    boolean deleteFeatures = featureRepository.batchDeleteByFeatureId(featureIds);
+    if (CollectionUtils.isEmpty(executePoints)) {
+      return deleteFeatures && testCaseRepository.deleteTestCase(caseId);
+    }
+
+    boolean deletePoints = executePointRepository.deleteByFeatureIds(featureIds);
+    return deleteFeatures && deletePoints && testCaseRepository.deleteTestCase(caseId);
+  }
+
+  public List<TestCaseDto> getE2ECases() {
+    return testCaseRepository.getE2ECases();
   }
 }
